show result
There will be three textboxes where I can enter numbers. I will enter numbers into these textboxes. When I press the "Show Result" button, the result of multiplying the sum of the numbers entered in the first two textboxes with the number entered in the third textbox will appear in a fourth textbox.